+++ docs/CHANGELOG.md
@@ -120,6 +120,7 @@ Semantic Versioning.
     assertion failure.
   * Subnamespaces can now be named contextual keywords such as `string`.
   * Import aliases can now be named contextual keywords such as `implements`.
+  * Import aliases can now be declared with `import type`.
   * Namespace aliases can now reference variables named contextual keywords such as
     `yield` inside namespaces.
   * Type annotations can now reference types inside namespaces named contextual